validation file for user-router
