+++ messaging/message-router.js
@@ -1,11 +1,4 @@
 const router = require("express").Router();
-const { check, validationResult, body } = require("express-validator");
-const Messages = require("./message-model.js");
-const MessageInbox = require("./message-inbox-model.js");
-const MessageReply = require("./message-reply-model");
-const Users = require("../users/user-model");
-const checkRole = require("../check-role/check-role-message.js");
-const restricted = require("../auth/restricted");
 
 const {
   validateMessageSend,
